Added new jahia:buildFrameworkPackageList goal that will go through the WEB-INF/classes and WEB-INF/lib directories to scan for packages and try to resolve their version. Currently split packages (= packages that exist in two directories) are handled by counting the number of entries in the packages and electing the one that has the most content but this is not always correct and something the version is guessed false. Another solution would be to export the package with all possible versions if that is allowed by OSGi (it seems so).
For the moment the goal outputs the list of packages in the console, and also sets a property called ${jahiaGeneratedFrameworkPackageList} that we could re-use in our Maven projects for example by setting up filtering in the felix-framework.properties file. This is relatively easy to do in the jahia-war project, but I'm not sure if we can do the same easily in the jahia-ee-war project, which we will need to do for the EE package.

There is also code that is using a previously generated MANIFEST.MF as input for the package list, assuming it is built using the Maven Bundle Plugin before calling this one. This has not been extensively tested and might be dropped since it might not be needed and a bit more complex to setup in a project.

The unit test only tests the MANIFEST.MF so it should probably be removed if we remove this functionality.



git-svn-id: file:///home/svn/repository/maven-jahia-plugin/trunk@626 2b131f5d-00a7-48c4-9630-96da95fee753